refactor: validate_serial_numbers
